requirements.txt
pillow
opencv-python
pyautogui
pytesseract
pathfinding
eventhandler
google-cloud-vision
shapely